ndi: unpackPhase without interpolation of src
The `unpackPhase` as of the previous commit did interpolate the result of the `src` on a nix level.
However for any overrides (`overrideAttrs`) to work on the source, for instance to update to a newer version locally, the `src` environment variable of `stdenv` must be used.
Otherwise the old, upstream, version will still be pulled in as a dependency and ultimately used for building, which will fail if the old version is not present, or will just deploy the old version if the old version was already downloaded (`requireFile` causes special semi-non-idempotent issues here).

Further information was provided in a GitHub comment: https://github.com/NixOS/nixpkgs/issues/219578#issuecomment-1586322972

Signed-off-by: benaryorg <binary@benary.org>